Fix for when reveal views have clickable control
Previously if we added a button with its `setOnClickListener` it will crash if we attempt to swipe (left or right) by starting with our finger directly over the button. Although the button is "hidden" behind the "front" view.

So this fix is mainly to keep both the reveal panels invisible (`GONE`) when then "front" view is covering them completely. I think this is probably what most developers would expect the default behavior to be.